+++ examples/capsulenet.py
@@ -16,10 +16,10 @@ Result:
 Author: Xifeng Guo, E-mail: `guoxifeng1990@163.com`, Github: `https://github.com/XifengGuo/CapsNet-Keras`
 """
 
-from keras import layers, models, optimizers
+from keras_capsules.capsulelayers import CapsuleLayer, PrimaryCap, Length, Mask
 from keras import backend as K
+from keras import layers, models, optimizers
 from keras.utils import to_categorical
-from capsulelayers import CapsuleLayer, PrimaryCap, Length, Mask
 
 
 def CapsNet(input_shape, n_class, num_routing):
