+++ setup.py
@@ -1,0 +1,16 @@
+from distutils.core import setup
+
+setup(
+    name='CapsNet-Keras',
+    version='',
+    packages=['keras_capsules'],
+    url='',
+    license='',
+    author='',
+    author_email='',
+    description='',
+    requires=[
+        'tensorflow',
+        'keras'
+    ]
+)
